+++ ai/langchain/src/main/kotlin/org/http4k/connect/langchain/chat/LmStudioChatLanguageModel.kt
@@ -49,7 +49,7 @@ data class LmStudioChatModelOptions(
     val temperature: Double = 1.0,
     val top_p: Double = 1.0,
     val n: Int = 1,
-    val stop: Any? = null,
+    val stop: List<String>? = null,
     val presencePenalty: Double = 0.0,
     val frequencyPenalty: Double = 0.0,
     val logitBias: Map<TokenId, Double>? = null,