+++ ai/lmstudio/client/src/main/kotlin/org/http4k/connect/lmstudio/action/ChatCompletion.kt
@@ -35,7 +35,7 @@ data class ChatCompletion(
     val temperature: Double = 1.0,
     val top_p: Double = 1.0,
     val n: Int = 1,
-    val stop: Any? = null,
+    val stop: List<String>? = null,
     val presence_penalty: Double = 0.0,
     val frequency_penalty: Double = 0.0,
     val logit_bias: Map<TokenId, Double>? = null,
@@ -156,10 +156,13 @@ data class Choice(
 
 @JsonSerializable
 data class ChoiceDetail(
-    val role: Role? = null,
+    @JsonProperty(name = "role")
+    internal val r: Role?,
     val content: String? = null,
     val tool_calls: List<ToolCall>? = null,
-)
+) {
+    val role = r ?: Role.Assistant
+}
 
 @JsonSerializable
 data class ToolCall(
@@ -194,7 +197,7 @@ data class CompletionResponse(
     val id: CompletionId,
     val created: Timestamp,
     val model: ModelName,
-    val choices: List<Choice>? = null,
+    val choices: List<Choice>,
     @JsonProperty(name = "object")
     val objectType: ObjectType,
     val usage: Usage? = null,