+++ ai/ollama/client/src/main/kotlin/org/http4k/connect/ollama/OllamaMoshi.kt
@@ -4,6 +4,7 @@ import com.squareup.moshi.JsonAdapter
 import com.squareup.moshi.Moshi
 import org.http4k.connect.model.Base64Blob
 import org.http4k.connect.model.ModelName
+import org.http4k.connect.model.Role
 import org.http4k.format.ConfigurableMoshi
 import org.http4k.format.ListAdapter
 import org.http4k.format.MapAdapter
@@ -21,6 +22,7 @@ object OllamaMoshi : ConfigurableMoshi(
         .value(Base64Blob)
         .value(ModelName)
         .value(Prompt)
+        .value(Role)
         .value(SystemMessage)
         .value(Template)
         .withStandardMappings()