+++ ai/openai/client/src/main/kotlin/org/http4k/connect/openai/action/ChatCompletion.kt
@@ -38,7 +38,7 @@ data class ChatCompletion(
     val temperature: Double = 1.0,
     val top_p: Double = 1.0,
     val n: Int = 1,
-    val stop: Any? = null,
+    val stop: List<String>? = null,
     val presence_penalty: Double = 0.0,
     val frequency_penalty: Double = 0.0,
     val logit_bias: Map<TokenId, Double>? = null,
@@ -48,6 +48,9 @@ data class ChatCompletion(
     val tools: List<Tool>? = null,
     val tool_choice: Any? = null,
     val parallel_tool_calls: Boolean? = null,
+    val service_tier: String? = null,
+    val seed: Int? = null,
+    val stream_options: StreamOptions? = null
 ) : OpenAIAction<Sequence<CompletionResponse>> {
     constructor(model: ModelName, messages: List<Message>, max_tokens: Int = 16, stream: Boolean = true) : this(
         model,
@@ -90,6 +93,8 @@ data class ChatCompletion(
         toCompletionSequence(response, OpenAIMoshi, "data: ", "[DONE]")
 }
 
+@JsonSerializable
+data class StreamOptions(val include_usage: Boolean? = null)
 
 @JsonSerializable
 @Polymorphic("type")
@@ -184,10 +189,13 @@ data class Choice(
 
 @JsonSerializable
 data class ChoiceDetail(
-    val role: Role? = null,
+    @JsonProperty(name = "role")
+    internal val r: Role?,
     val content: String? = null,
     val tool_calls: List<ToolCall>? = null,
-)
+) {
+    val role = r ?: Role.Assistant
+}
 
 @JsonSerializable
 data class ToolCall(
@@ -222,8 +230,10 @@ data class CompletionResponse(
     val id: CompletionId,
     val created: Timestamp,
     val model: ModelName,
-    val choices: List<Choice>? = null,
+    val choices: List<Choice>,
     @JsonProperty(name = "object")
     val objectType: ObjectType,
     val usage: Usage? = null,
+    val system_fingerprint: String? = null,
+    val service_tier: String? = null
 )