+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 This list is not intended to be all-encompassing - it will document major and breaking API 
 changes with their rationale when appropriate. Given version `A.B.C.D`, breaking changes are to be expected in version number increments where changes in the `A` or `B` sections:
 
+### v5.24.1.0 (uncut)
+- **http4k-connect-ai-** - Chat complection
+
 ### v5.24.0.0
 - **http4k-connect-*** - Upgrade dependencies including
 - **http4k-connect-ai-openai*** - [Breaking] Tightened up types for completion requests.